uio: split DiskCacheFs into its own module
