+++ lib/common/common/src/universal_io/mod.rs
@@ -15,7 +15,7 @@ pub use self::error::UniversalIoError;
 #[cfg(target_os = "linux")]
 pub use self::io_uring::{IoUringFile, IoUringFs, IoUringOpenExtra};
 pub use self::mmap::{MmapFile, MmapFs};
-pub use self::simple_disk_cache::{DiskCache, DiskCacheConfig};
+pub use self::simple_disk_cache::{DiskCache, DiskCacheConfig, DiskCacheFs};
 pub use self::traits::{
     BorrowedReadPipeline, Item, OwnedReadPipeline, TConfigContext, UniversalRead,
     UniversalReadFileOps, UniversalReadFs, UniversalWrite, UserData,
